reorganize dto and mapper folders
