+++ src/main/java/io/github/codenilson/lavava2025/entities/dto/player/PlayerUpdateDTO.java
@@ -1,4 +1,4 @@
-package io.github.codenilson.lavava2025.dto.player;
+package io.github.codenilson.lavava2025.entities.dto.player;
 
 import lombok.Data;
