+++ src/test/java/io/github/codenilson/lavava2025/unit/services/PlayerServiceTest.java
@@ -25,14 +25,14 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
-import io.github.codenilson.lavava2025.dto.player.PlayerCreateDTO;
-import io.github.codenilson.lavava2025.dto.player.PlayerResponseDTO;
-import io.github.codenilson.lavava2025.dto.player.PlayerUpdateDTO;
 import io.github.codenilson.lavava2025.entities.Player;
+import io.github.codenilson.lavava2025.entities.dto.player.PlayerCreateDTO;
+import io.github.codenilson.lavava2025.entities.dto.player.PlayerResponseDTO;
+import io.github.codenilson.lavava2025.entities.dto.player.PlayerUpdateDTO;
+import io.github.codenilson.lavava2025.entities.mappers.PlayerMapper;
 import io.github.codenilson.lavava2025.entities.valueobjects.Roles;
 import io.github.codenilson.lavava2025.errors.PlayerNotFoundException;
 import io.github.codenilson.lavava2025.errors.UsernameAlreadyExistsException;
-import io.github.codenilson.lavava2025.mappers.PlayerMapper;
 import io.github.codenilson.lavava2025.repositories.PlayerRepository;
 import io.github.codenilson.lavava2025.services.PlayerService;
